backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Streamlit para frontend
streamlit==1.39.0

# Azure Services
openai==1.54.3
azure-ai-documentintelligence==1.0.0b4
azure-core==1.31.0

# Modelos de datos
pydantic==2.9.0
pydantic-settings==2.5.0

# Utilidades
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.24.0

# Procesamiento de archivos
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==10.4.0

# Testing
pytest==8.3.0
httpx==0.27.2
